renaming from details to map, adding flag for whether or not to write out the map json file
